Update instrumentation subpackage structure (#284)
