+++ nevergrad/__init__.py
@@ -5,7 +5,7 @@
 
 from .optimization import optimizerlib as optimizers  # busy namespace, likely to be simplified
 from .optimization import families
-from .instrumentation import Instrumentation
+from .instrumentation.multivariables import Instrumentation
 from .instrumentation import variables as var
 
 __all__ = ["Instrumentation", "var", "optimizers", "families"]
