+++ nevergrad/instrumentation/core.py
@@ -1,286 +1,155 @@
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved.
-#
-# This source code is licensed under the MIT license found in the
-# LICENSE file in the root directory of this source tree.
-
-import itertools
-from typing import List, Any, Tuple, Dict, Optional, Callable
+import copy
+from typing import Any, Tuple, Optional, Dict, Set, TypeVar
 import numpy as np
 from ..common.typetools import ArrayLike
-from . import utils
-from . import variables
-from .core2 import Variable, ArgsKwargs
+
+ArgsKwargs = Tuple[Tuple[Any, ...], Dict[str, Any]]
+T = TypeVar('T', bound="Variable")
 
 
-class Instrumentation(Variable):
-    """Class handling arguments instrumentation, and providing conversion to and from data.
+class VarSpecs:
 
-    Parameters
-    ----------
-    *args:
-        Any positional argument. Arguments of type Variable (see note) will serve for instrumentation
-        and others will be kept constant.
-    **kwargs
-        Any keyword argument. Arguments of type Variable (see note) will serve for instrumentation
-        and others will be kept constant.
+    # pylint: disable=too-many-arguments, unused-argument
+    def __init__(self) -> None:
+        self.dimension = -1
+        self.nargs = 1
+        self.kwargs_keys: Set[str] = set()
+        self.continuous = True
+        self.noisy = False
 
-    Note
-    ----
-    * Variable classes are:
+    def update(self,
+               dimension: Optional[int] = None,
+               nargs: Optional[int] = None,
+               kwargs_keys: Optional[Set[str]] = None,
+               continuous: Optional[bool] = None,
+               noisy: Optional[bool] = None
+               ) -> None:
+        for key, value in locals().items():
+            if key != "self" and value is not None:
+                setattr(self, key, value)
 
-        - `SoftmaxCategorical(items)`: converts a list of `n` (unordered) categorial items into an `n`-dimensional space. The returned
-          element will be sampled as the softmax of the values on these dimensions. Be cautious: this process is non-deterministic
-          and makes the function evaluation noisy.
-        - `OrderedDiscrete(items)`: converts a list of (ordered) discrete items into a 1-dimensional variable. The returned value will
-          depend on the value on this dimension: low values corresponding to first elements of the list, and high values to the last.
-        - `Gaussian(mean, std)`: normalizes a `n`-dimensional variable with independent Gaussian priors (1-dimension per value).
-        - `Array(dim1, ...)`: casts the data from the optimization space into a np.ndarray of any shape,
-          to which some transforms can be applied (see `asscalar`, `affined`, `exponentiated`, `bounded`).
-          This is therefore a very flexible type of variable.
-        - `Scalar(dtype)`: casts the data from the optimization space into a float or an int. It is equivalent to `Array(1).asscalar(dtype)`
-          and all `Array` methods are therefore available
-    * Depending on the variables, `Instrumentation` can be noisy (`SoftmaxCategorical` in non-deterministic mode), and not continuous
-      (`SoftmaxCategorical` in deterministic mode, `OrderedDiscrete`, `Array` with int casting). Some optimizers may not be able
-      to deal with these cases properly.
-    * Since instrumentation can have a stochastic behavior, it is responsible of holding the random state which optimizers
-      pull from. This random state is initialized lazily, and can be modified
-    """
 
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__()
-        self.names: Tuple[Optional[str], ...] = ()
-        self.variables: List[Variable] = []
-        self._set_args_kwargs(args, kwargs)
+class Variable:
+
+    def __init__(self) -> None:
         self._name: Optional[str] = None
         self._random_state: Optional[np.random.RandomState] = None  # lazy initialization
+        self._specs = VarSpecs()
 
     @property
-    def args(self) -> Tuple[Variable, ...]:
-        """List of variables passed as positional arguments
+    def random_state(self) -> np.random.RandomState:
+        """Random state the instrumentation and the optimizers pull from.
+        It can be seeded/replaced.
         """
-        return tuple(arg for name, arg in zip(self.names, self.variables) if name is None)
+        if self._random_state is None:
+            # use the setter, to make sure the random state is propagated to the variables
+            seed = np.random.randint(2 ** 32, dtype=np.uint32)
+            self._set_random_state(np.random.RandomState(seed))
+        assert self._random_state is not None
+        return self._random_state
 
-    @property
-    def kwargs(self) -> Dict[str, Variable]:
-        """Dictionary of variables passed as named arguments
-        """
-        return {name: arg for name, arg in zip(self.names, self.variables) if name is not None}
-
-    def _set_args_kwargs(self, args: Tuple[Any, ...], kwargs: Dict[str, Any]) -> None:
-        self.names, arguments = self._make_argument_names_and_list(args, kwargs)
-        self.variables = [variables._Constant.convert_non_instrument(a) for a in arguments]
-        assert all(v.nargs == 1 and not v.kwargs_keys for v in self.variables), "Not yet supported"
-        num_instru = len(set(id(i) for i in self.variables))
-        assert len(self.variables) == num_instru, "All instruments must be different (sharing is not supported)"
-        self._specs.update(
-            dimension=sum(i.dimension for i in self.variables),
-            continuous=all(v.continuous for v in self.variables),
-            noisy=any(v.noisy for v in self.variables),
-            nargs=len(args),
-            kwargs_keys=set(kwargs.keys()),
-        )
+    @random_state.setter
+    def random_state(self, random_state: np.random.RandomState) -> None:
+        self._set_random_state(random_state)
 
     def _set_random_state(self, random_state: np.random.RandomState) -> None:
-        super()._set_random_state(random_state)
-        assert self._random_state is not None
-        if self.variables:
-            for var in self.variables:
-                var._random_state = self._random_state
+        self._random_state = random_state
 
-    @staticmethod
-    def _make_argument_names_and_list(args: Tuple[Any, ...], kwargs: Dict[str, Any]) -> Tuple[Tuple[Optional[str], ...], Tuple[Any, ...]]:
-        """Converts *args and **kwargs to a tuple of names (with None for positional),
-        and the corresponding tuple of values.
-
-        Eg:
-        _make_argument_names_and_list(3, z="blublu", machin="truc")
-        >>> (None, "machin", "z"), (3, "truc", "blublu")
+    def with_name(self: T, name: str) -> T:
+        """Sets a name and return the current instrumentation (for chaining)
         """
-        names: Tuple[Optional[str], ...] = tuple([None] * len(args) + sorted(kwargs))  # type: ignore
-        arguments: Tuple[Any, ...] = args + tuple(kwargs[x] for x in names if x is not None)
-        return names, arguments
+        self._name = name
+        return self
 
-    def _data_to_arguments(self, data: np.ndarray, deterministic: bool = True) -> ArgsKwargs:
-        arguments = utils.process_variables(self.variables, data, deterministic=deterministic)
-        args = tuple(arg for name, arg in zip(self.names, arguments) if name is None)
-        kwargs = {name: arg for name, arg in zip(self.names, arguments) if name is not None}
-        return args, kwargs
+    @property
+    def name(self) -> str:
+        if self._name is not None:
+            return self._name
+        return format(self, "short")
+
+    def copy(self: T) -> T:  # TODO, use deepcopy directly in the code if it works?
+        """Return a new instrumentation with the same variable and same name
+        (but a different random state)
+        """
+        instru = copy.deepcopy(self)
+        instru._random_state = None
+        return instru
+
+    @property
+    def dimension(self) -> int:
+        return self._specs.dimension
+
+    @property
+    def nargs(self) -> int:
+        return self._specs.nargs
+
+    @property
+    def kwargs_keys(self) -> Set[str]:
+        return self._specs.kwargs_keys
+
+    @property
+    def continuous(self) -> bool:
+        return self._specs.continuous
+
+    @property
+    def noisy(self) -> bool:
+        return self._specs.noisy
+
+    def arguments_to_data(self, *args: Any, **kwargs: Any) -> np.ndarray:
+        """Converts args and kwargs into data in np.ndarray format
+        """
+        if len(args) != self.nargs:
+            raise TypeError(f"Expected {self.nargs} arguments ({len(args)} given: {args})")
+        if self.kwargs_keys != set(kwargs.keys()):
+            raise TypeError(f"Expected arguments {self.kwargs_keys} ({set(kwargs.keys())} given: {kwargs})")
+        return self._arguments_to_data(*args, **kwargs)
 
     def _arguments_to_data(self, *args: Any, **kwargs: Any) -> np.ndarray:
-        """Converts arguments to data
+        raise RuntimeError(f"arguments_to_data is not defined for {self.__class__.__name__}")
 
-        Parameters
-        ----------
-        *args: Any
-            the positional arguments corresponding to the instance initialization positional arguments
-        **kwargs: Any
-            the keyword arguments corresponding to the instance initialization keyword arguments
-
-        Returns
-        -------
-        data: np.ndarray
-            the corresponding data in the optimization space
-
-        Note
-        ----
-        - you need to input the arguments in the same way than at initialization
-          with regard to positional and named arguments.
-        - this process is simplified, and is deterministic. Depending on your instrumentation,
-          you will probably not recover the same data.
-        """
-        names, arguments = self._make_argument_names_and_list(args, kwargs)
-        assert names == self.names, (f"Passed argument pattern (positional Vs named) was:\n{names}\n"
-                                     f"but expected:\n{self.names}")
-        data = list(itertools.chain.from_iterable([instrument.arguments_to_data(arg)
-                                                   for instrument, arg in zip(self.variables, arguments)]))
-        return np.array(data)
-
-    def instrument(self, function: Callable[..., Any]) -> "InstrumentedFunction":
-        return InstrumentedFunction(function, *self.args, **self.kwargs)
-
-    def split_data(self, data: ArrayLike) -> List[np.ndarray]:
-        """Splits the input data in chunks corresponding to each of the variables in self.variables
+    def data_to_arguments(self, data: ArrayLike, deterministic: bool = False) -> ArgsKwargs:
+        """Converts data to arguments
 
         Parameters
         ----------
         data: ArrayLike (list/tuple of floats, np.ndarray)
             the data in the optimization space
+        deterministic: bool
+            whether the conversion should be deterministic (some variables can be stochastic, if deterministic=True
+            the most likely output will be used)
 
         Returns
         -------
-        List[np.ndarray]
-            the list of data chunks corresponding to each variable in self.variables
+        args: Tuple[Any]
+            the positional arguments corresponding to the instance initialization positional arguments
+        kwargs: Dict[str, Any]
+            the keyword arguments corresponding to the instance initialization keyword arguments
         """
-        return utils.split_data(data, self.variables)
+        # trigger random_state creation (may require to be propagated to sub-variables
+        assert self.random_state is not None
+        array = np.array(data, copy=False)
+        if array.shape != (self.dimension,):
+            raise ValueError(f"Unexpected shape {array.shape} of {array} for {self} with dimension {self.dimension}")
+        return self._data_to_arguments(array, deterministic)
+
+    def _data_to_arguments(self, data: np.ndarray, deterministic: bool) -> ArgsKwargs:
+        raise NotImplementedError
+
+    def get_summary(self, data: np.ndarray) -> str:  # pylint: disable=unused-argument
+        return f"No configured summary in {self}"
+
+    # def __eq__(self, other: Any) -> bool:
+    #    return bool(self.__class__ == other.__class__ and self.__dict__ == other.__dict__)
+
+    def _short_repr(self) -> str:
+        return repr(self)
 
     def __format__(self, format_spec: str) -> str:
-        arguments = [format(x, format_spec) for x in self.args]
-        sorted_kwargs = [(name, format(self.kwargs[name], format_spec)) for name in sorted(self.kwargs)]
-        all_params = arguments + [f"{x}={y}" for x, y in sorted_kwargs]
+        if self._name is not None:
+            return self._name
         if format_spec == "short":
-            return ",".join(all_params)
-        return "{}({})".format(self.__class__.__name__, ", ".join(all_params))
-
-    def __repr__(self) -> str:
-        return f"{self:display}"
-
-    def get_summary(self, data: ArrayLike) -> Any:
-        """Provides the summary string corresponding to the provided data
-
-        Note
-        ----
-        This is impractical for large arrays
-        """
-        strings = []
-        splitted_data = utils.split_data(data, self.variables)
-        for k, (name, var, d) in enumerate(zip(self.names, self.variables, splitted_data)):
-            if not isinstance(var, variables._Constant):
-                explanation = var.get_summary(d)
-                sname = f"arg #{k + 1}" if name is None else f'kwarg "{name}"'
-                strings.append(f"{sname}: {explanation}")
-        return " - " + "\n - ".join(strings)
-
-
-class InstrumentedFunction:
-    """Converts a multi-argument function into a single-argument multidimensional continuous function
-    which can be optimized.
-
-    Parameters
-    ----------
-    function: callable
-        the callable to convert
-    *args, **kwargs: Any
-        Any argument. Arguments of type Variable (see notes) will be instrumented,
-        and others will be kept constant.
-
-    Notes
-    -----
-    - Variable classes are:
-        - `SoftmaxCategorical(items)`: converts a list of `n` (unordered) categorial items into an `n`-dimensional space. The returned
-           element will be sampled as the softmax of the values on these dimensions. Be cautious: this process is non-deterministic
-           and makes the function evaluation noisy.
-        - `OrderedDiscrete(items)`: converts a list of (ordered) discrete items into a 1-dimensional variable. The returned value will
-           depend on the value on this dimension: low values corresponding to first elements of the list, and high values to the last.
-        - `Gaussian(mean, std)`: normalizes a `n`-dimensional variable with independent Gaussian priors (1-dimension per value).
-        - `Array(dim1, ...)`: casts the data from the optimization space into a np.ndarray of any shape,
-          to which some transforms can be applied (see `asscalar`, `affined`, `exponentiated`, `bounded`).
-          This is therefore a very flexible type of variable.
-        - `Scalar(dtype)`: casts the data from the optimization space into a float or an int. It is equivalent to `Array(1).asscalar(dtype)`
-          and all `Array` methods are therefore available
-    - This function can then be directly used in benchmarks *if it returns a float*.
-    - You can update the "_descriptors" dict attribute so that function parameterization is recorded during benchmark
-    """
-
-    def __init__(self, function: Callable[..., Any], *args: Any, **kwargs: Any) -> None:
-        assert callable(function)
-        self._descriptors: Dict[str, Any] = {"function_class": self.__class__.__name__}
-        self._instrumentation = Instrumentation()  # dummy
-        self.instrumentation = Instrumentation(*args, **kwargs)  # sets descriptors
-        self.function = function
-        self.last_call_args: Optional[Tuple[Any, ...]] = None
-        self.last_call_kwargs: Optional[Dict[str, Any]] = None
-        # if this is not a function bound to this very instance, add the function/callable name to the descriptors
-        if not hasattr(function, '__self__') or function.__self__ != self:  # type: ignore
-            name = function.__name__ if hasattr(function, "__name__") else function.__class__.__name__
-            self._descriptors.update(name=name)
-
-    @property
-    def instrumentation(self) -> Instrumentation:
-        return self._instrumentation
-
-    @instrumentation.setter
-    def instrumentation(self, instrum: Instrumentation) -> None:
-        self._instrumentation = instrum
-        self._descriptors.update(instrumentation=instrum.name, dimension=instrum.dimension)
-
-    @property
-    def dimension(self) -> int:
-        return self.instrumentation.dimension
-
-    def data_to_arguments(self, data: ArrayLike, deterministic: bool = True) -> Tuple[Tuple[Any, ...], Dict[str, Any]]:
-        """Get the arguments and keyword arguments corresponding to the data
-
-        Parameters
-        ----------
-        data: np.ndarray
-            input data
-        deterministic: bool
-            whether to process the data deterministically (some Variables such as SoftmaxCategorical are stochastic).
-            If True, the output is the most likely output.
-        """
-        return self.instrumentation.data_to_arguments(data, deterministic=deterministic)
-
-    def arguments_to_data(self, *args: Any, **kwargs: Any) -> ArrayLike:
-        return self.instrumentation.arguments_to_data(*args, **kwargs)
-
-    def __call__(self, x: ArrayLike) -> Any:
-        self.last_call_args, self.last_call_kwargs = self.data_to_arguments(x, deterministic=False)
-        return self.function(*self.last_call_args, **self.last_call_kwargs)
-
-    def get_summary(self, data: ArrayLike) -> Any:  # probably impractical for large arrays
-        """Provides the summary corresponding to the provided data
-        """
-        return self.instrumentation.get_summary(data)
-
-    @property
-    def descriptors(self) -> Dict[str, Any]:
-        """Description of the function parameterization, as a dict. This base class implementation provides function_class,
-            noise_level, transform and dimension
-        """
-        return dict(self._descriptors)  # Avoid external modification
-
-    def __repr__(self) -> str:
-        """Shows the function name and its summary
-        """
-        params = [f"{x}={repr(y)}" for x, y in sorted(self._descriptors.items())]
-        return "Instance of {}({})".format(self.__class__.__name__, ", ".join(params))
-
-    def __eq__(self, other: Any) -> bool:
-        """Check that two instances where initialized with same settings.
-        This is not meant to be used to check if functions are exactly equal (initialization may hold some randomness)
-        This is only useful for unit testing.
-        (may need to be overloaded to make faster if tests are getting slow)
-        """
-        if other.__class__ != self.__class__:
-            return False
-        return bool(self._descriptors == other._descriptors)
+            return self._short_repr()
+        elif format_spec == "display":
+            # ugly hack below, but simplifies code a lot
+            return self._short_repr() if self.__class__.__name__ == "_Constant" else repr(self)
+        return repr(self)
