+++ nevergrad/optimization/optimizerlib.py
@@ -311,9 +311,8 @@ class EDA(base.Optimizer):
             self.evaluated_population_sigma = [p[1] for p in sorted_pop_with_sigma_and_fitness]
             self.evaluated_population_fitness = [p[2] for p in sorted_pop_with_sigma_and_fitness]
             # Computing the new parent.
-            self.current_center = (  # type: ignore
-                sum([np.asarray(self.evaluated_population[i]) for i in range(self.mu)]) / self.mu
-            )
+            arrays = [np.asarray(self.evaluated_population[i]) for i in range(self.mu)]
+            self.current_center = sum(arrays) / self.mu  # type: ignore
             self.sigma = np.exp(sum([np.log(self.evaluated_population_sigma[i]) for i in range(self.mu)]) / self.mu)
             self.evaluated_population = []
             self.evaluated_population_sigma = []
@@ -371,9 +370,8 @@ class PCEDA(EDA):
             self.evaluated_population_sigma = [p[1] for p in sorted_pop_with_sigma_and_fitness]
             self.evaluated_population_fitness = [p[2] for p in sorted_pop_with_sigma_and_fitness]
             # Computing the new parent.
-            self.current_center = (  # type: ignore
-                sum([np.asarray(self.evaluated_population[i]) for i in range(self.mu)]) / self.mu
-            )
+            arrays = [np.asarray(self.evaluated_population[i]) for i in range(self.mu)]
+            self.current_center = sum(arrays) / self.mu  # type: ignore
             self.sigma = np.exp(sum([np.log(self.evaluated_population_sigma[i]) for i in range(self.mu)]) / self.mu)
             self.evaluated_population = []
             self.evaluated_population_sigma = []
@@ -429,9 +427,8 @@ class MPCEDA(EDA):
             self.evaluated_population_sigma = [p[1] for p in sorted_pop_with_sigma_and_fitness]
             self.evaluated_population_fitness = [p[2] for p in sorted_pop_with_sigma_and_fitness]
             # Computing the new parent.
-            self.current_center = (  # type: ignore
-                sum([np.asarray(self.evaluated_population[i]) for i in range(self.mu)]) / self.mu
-            )
+            arrays = [np.asarray(self.evaluated_population[i]) for i in range(self.mu)]
+            self.current_center = sum(arrays) / self.mu  # type: ignore
             self.sigma = np.exp(sum([np.log(self.evaluated_population_sigma[i]) for i in range(self.mu)]) / self.mu)
             self.evaluated_population = []
             self.evaluated_population_sigma = []
@@ -467,9 +464,8 @@ class MEDA(EDA):
             self.evaluated_population_sigma = [p[1] for p in sorted_pop_with_sigma_and_fitness]
             self.evaluated_population_fitness = [p[2] for p in sorted_pop_with_sigma_and_fitness]
             # Computing the new parent.
-            self.current_center = (  # type: ignore
-                sum([np.asarray(self.evaluated_population[i]) for i in range(self.mu)]) / self.mu
-            )
+            arrays = [np.asarray(self.evaluated_population[i]) for i in range(self.mu)]
+            self.current_center = sum(arrays) / self.mu  # type: ignore
             self.sigma = np.exp(sum([np.log(self.evaluated_population_sigma[i]) for i in range(self.mu)]) / self.mu)
             self.evaluated_population = []
             self.evaluated_population_sigma = []
